mover parser.go -> listsecretparser.go
We'll be introducing a new parser shortly, better to have these in
seperate files.
